refactor: move test builders to separate modules
